Add basic dvektor
This is like vektor, but uses a *focus* and *display*, like the Scala
immutable vector implementation.  This display keeps a path to a
specific chunk of the vector out of the tree.  This allows faster access
around the last location that was modified.

We removed the loop unrolling that the Scala implementation does, to
make benchmark results more comparable to our other Clojure-inspired
implementation.  More experiments are needed to arrive at conclusions on
which is better.