Add custom tagged union for vektor nodes
This way we do not pay the cost of type checks when casting it, since we
are always sure about the type -- the type is determined by the location
in the data structure.

Note that a tag is still needed for the destructor call, since we rely
on `shared_ptr` for doing the reference counting, which does not know
about the position of the node in the tree.  We might do reference
counting manually in the future in the data structure itself, so
implicit knowledge about the node type in the traversal is not erased,
but that is an optimization that we'll do later since it increases
complexity sustantially.