Properly manage vektor element instances
This commit allows fixes management of element values in a `vektor`.
Before, default construction was required because we were always
allocating full `array<>`s with `1<<B` elements, even when not all
objects where actually added.  Now, default construction is no longer
required.

Also, since we needed to touch the allocator API for our free list
optimization, we have improved it to allow generic heap composition in
the spirit of the `Heap Layers` library [1].

[1] https://github.com/emeryberger/Heap-Layers