Fix crashes when running benchmarks for GC-powered vektor
The problem was that we were capturing _vektors_ in a lambda a returning
it to the benchmark, which would in turn it store it in the heap inside
of `nonius::benchmark_function`.  Since `libgc` can not see inside memory
allocated with `malloc`, this was causing the `vektor` nodes to be
prematurelly garbage collected.