Sketch RRB vector operations
Note that not all of these operations have actually been tested, since
some of the states can not be reached because concatenation has not been
fully implemented.

Concatenation is, btw, broken in this commit.  It only works when the
right hand sides has `<= branches<B>` elements, in which case we just
grab its tail.