Fix push_tail_into_root() was not dealing properly with overflow
It would sometimes fail to push the tail into a relaxed vector.  It
would try to push the tail in a branch that overflows causing to
either (1) create node with mora than `branches<B>` slots, or (2)
overwrite some slot with the tail.  All that is bad.

Now we remove our incomplete attempt of detecting overflow for relaxed
trees in `is_overflow` and submit to just backtrack if our leftmost
branch did not have space for the tail.  This might be even faster -- we
only navigate the tree once.