Do not store slot counts in regular subtrees
In a regular tree, the slot count of a node can be derived from the size
of the tree and the position of the node.  We already leverage this in
the `vektor` type, but until this change we did store the slot count in
regular `rvektor` nodes as to traversal and concatenation easier.

Admittedly, the code is now a mess.

Also, this commit changes other things, like some naming conventions (e.g.
say `_regular` instead of `_last` to indicate visiting a regular tree
node that is not necessarily full) and some functions extracted (e.g
`update_leaf`) and code cleanup (e.g semantics of the `push_tail` that
replace `push_tail_into_root`).

Sorry for the mess, I will cleanup soon...