Do not store node type tags in release mode for relaxed vectors
They were only needed to be able to compute the depth of a tree.  We did
this because we were lazy when implementing `concat` and did not track
the new tree depth while building the concatenated tree.  Now we do so.

This means that there is no superfluous information in the relaxed
vektor nodes.  For regular trees, the only cost is having a pointer to
the sizes vector in inner nodes.  Most probably, we will later use a
single node representation for `vektor` and `rvektor`, this allowing
trivial conversion from `vektor` to `rvektor`.  This might allow the
user to only pay for the cost of an `rvektor` iff at some point they
need to concatenate it.