Rename rvektor to flex_vector
In the spirit of standard data structures, it is better to name it
refering to its observable properties than to the academic naming.  In
this sense, maybe it is better to say that this is a "flexible vector",
i.e. something vector-like but more rope-like, than to say it is a
relaxed vector, which refers to the internal invariants of the
implementation tree structure.