Rename ivektor to array
Again, this is a less alien name, and a bit more descriptive.  Unlike
`std::vectors`, this is not compile-time sized.  Since inmutable data
structures are heap-allocated anyways, there is no advantage to having
it constant-sized.

However, this data structure, unlike `vector`, has trully sequential
access and thus also `O(n)` updates...

It needs a bit more work though to catch up with the design of the other
structures.