Extract r/rb operations into generic traversal
Using a continuation passing style of visitation, we can traverse the
trees encoding the position in a type.  This type can implement the most
efficient algorithms for this node in this context of the tree.

This allows to write certain operations in a generic way, we just
dispatch to the correct (regular or relaxed) kind of node at the root.

There seems to be an almost negible but existing performance penalty for
this approach, probably because the compilers have a harder time
collapsing these traversal objects than the parameters that we passed
before.  Or maybe because we went too far in specializing the traversal.
We can investigate that later, for now the approach is promising at
cleaning up the code and making optimizations easier to write.