Generalize slice_right for rb and rrb trees
For this, we needed to change slightly the API for building visitors,
such that it does not rely so much on lambda soups.  I know, the older
implementation was beauuuuutiful, but it had some problems:

1. It was driving Clang crazy when building recursive visitors causing
infinite template instantiation when (incidently, GCC was happy with
them if I just added a bit more type hinting, and I think the code was
correct.)

2. I was generating crazy name manglings that was causing GDB to crash
when loading the file. (This is LOL).

3. It was (slightly) harder to inline and optimize.  The new structure
seems to allow the compiler to generate better code in some cases.
Results were surprisingly good for the old code nonetheless.

Note to self: try to isolate problems 1) and 2) and submit the
appropriate bug reports.