Implement slice_left using generalized traversal
Note that this algorithm does not work for rbtrees though, since a
`slice_left` operation always produces a relaxed tree.

Also, I am not very happy with the current design which adds
`this_size()`.  Ideally, we should just compute this alongside the
traversal.  I would like to find a way to do this without adding an
overhead to other traversals that do not require the `this_size()`.

Furthermore, since in `slice_left` we always produce a relaxed node, I
would like to only use one kind of traversal step for all inner (relaxed
and otherwise) nodes.  This would also be a good step towards
implementing concatenation using the new facilities.