Have separate position hierarchies for exact and inexact sizes
These can be reached with different navigation functions.  The core idea
is that the `_sub` variations (e.g. `towards_sub`) adapt the index and
sizes so they are accurate for the subtree.  In contrast, the
alternative versions *might* return the size and indexes of parent
subtrees.

This allows us to gain back the few cycles that we lost in:

    eb6e9db Implement slice_left generically for relaxed and...