Write update_visitor as stateless visitor
This means we needed to adapt the `towards...` kind of traversal
functions to thread remaining arguments for the recursion.  This
required disambiguating the _hints_ by function name instead of by
positional overloads.  This is a good idead anyways, since there are
hints for the same traversal that have different meanings but fall in
the same position.  It would be easy to fall into traps.