Only allocate the exact amount of elements needed
This has a significant performance boost when using `libgc`.  It
is conservative garbage collector, so it also scans the allocated
"rests" of memory at the end of nodes that were not fully used.

There was a reason, though, why we were allocating full nodes
unconditionally.  When we add transients, we plan to introduce an
optimization where, when using garbage collection, we use mutating
operations on nodes with a single owner that come from an
r-value-reference.  When we add such feature, we will decide which
strategy for allocation to use depending on wether that optimization is
applicable.

NB. This change also required changing the layout of the nodes, such the
"variable length" sections is always at the end.  This may have a tiny
effect on the efficiency of cache utilization for RB-vectors, but our
measurements say it is negible.  The layout change may later be made
optional -- depending on the usage of reference counting.