Use libgc stubborn API
This API allows to mark objects that will not change anymore.

In theory, this can allow for faster collections because these objects
can be scanned just once.  This seems ideal for immutable data
structures!  For these, we can `freeze()` nodes.  Note that we changed
the implementation of some algorithms to avoid allocations that happen
while some node is still unfrozen.

In practice, this API is deprecated in libgc it has no effect -- too bad
I noticed only after implementing the change and measuring it the
performance change.  I leave the commit here to document the
exploration, will revert later.