Implement concatenation using positions and visitors...
    Everyone knows that debugging is twice as hard as writing a program
    in the first place. So if you are as clever as you can be when you
    write it, how will you ever debug it?

        -- Brian Kernighan

This commit leaves me filled with self doubt.

On the one hand, this encoding of traversal in generic alagorithms
applied to navigation through typed positions is interesting.  It makes
reading the actual algorithms tolerable (for some definition of...), to
the extent that there is not a lot of weird bit manipulation all over
just to navigate the tree.  Also, we manage to remove all kinds of
branching through the code, as once a property about about a whole
subtree is discovered, it is encoded in a type and dispatched statically
through the rest of the subtree.

However, there are several problems, like (1) gdb refuses, again, to
demangle the generated function names, in spite of me having got rid of
all the lambdas, (2) there is a potential for unadvertedly exploding the
generated code because of combinatorial explosion between traversal and
algorithmic states, (3) a smart compiler is required for this code to be
efficient, (4) without sufficient understanding of templates and high
order programming, this code must feel like magic, etc... (5) adding
exception safety (to come in the next commits) is gonna be non-trivial,
because a node does not know how to "delete itself", (6) etc...?

Most other implementations of RRB trees manage to be much simpler just
by storing the slot count per node and having some kind of per node type
tag.  With this information a node knows how to delete itself and all
kinds of regular subtrees can be traversed in simpler and more
homogeneous ways.  However, we are adding "superfluous" information to
the tree.  As a C++ programmer obsessed with squeezing every byte and
every cycle, we managed to prove the point: it is possible to implement
RRB-trees without any metadata in the nodes.  The question remains...

...is it worth it?