Fix BL == 0
In fact, this case can be covered by the general code paths, but an
assertion was needed to be removed.

This assertion was only useful to catch some errors in our initial
coding of the algorithm, but does not hold in the general case.  In
fact, as the failing test hilighted, it can be triggered when
concatenating a vector with itself.  This is easier to trigger now that
singleton leaves are possible.