Fix relaxed vector exception safety
I am not happy about the `catch` block in
`push_tail_visitor::visit_regular`, but at the same time avoiding this
complexity would require a significant re-thinking of the whole
ownership management, with potential performance loss.  This would mean,
as insinuated in the previous commit, adding a `position` wrapper that
does ownership management using RAII and passing these around all the
time instead of raw pointers with ad-hoc ownership semantics.  But this
might have some slight performance impact and add complexities of its
own -- since this is a write-once-use-often library it might be worth
paying the cost of complexity.  Let's think about this further in the
morning when my brain actually works...