Fix regular vector exception safety
Before, it would leak memory.

Note that we avoided using RAII to add exception safety.  This makes the
code arguably complicated and brittle.  But given that nodes don't know
how to destroy themselves (see rationale for the existence of
`positions`) it is not so easy to use RAII here.  It could be done by
leveraging positions (using a `position guard` or something).  But this
would add a few cycles here and because 1) destructors need to be called
for non-error code paths and do the whole `if (!moved)` thing and 2)
exceptional paths need to call the whole tree-traversing `dec`.  Doing
it manually we can be smart and do all allocations and copying of leaves
*first*, and all no-throw `inc()` and and copying later.

The way we check exception safety is also kinda... fun.  The idea is to
have some *absurd* --i.e. inintellible, random-looking-- process that is
though deterministic and manages to throws exceptions on various
interesting points.  This system is re-entrant but not thread safe
because of the way it uses global state, to make it easier to
communicate to the potentially throwing points.