Submit to using ReStructuredText
Since we need to fall back to it a lot of times anyways, let's just be
ok with it and use it as much as possible.  It is indeed a nice
language, some things are better than Markdown, some are worse.  We
leave Recommonmark there so people can still write Markdown if they
really hate RST.