Sketch transience policies
We need to know if a node is "modifiable" in place as part of a
transient manipulation operation.  We can do this in two ways:

1. When we use reference counting, we can just check if the object is
   "unique", i.e. there is only one reference to it.

2. If we don't have reference counts, we can do as in Clojure... when
   a node is created, we set the identity of its root node as
   "owner".  Operations on the owner can modify it in place.

There is a policy to enable or disable the approach 2.  The
`rbts::node` combines 1. and 2. by asking the refcount and transience
policies in the `can_modify` methods.

Note:
   In this commit we also change the refcount policies to work in a
   more OO-way... This was not really necessary, but we go driven away
   in the middle of the experimentation.  It is just a syntactic
   change, but it actually makes the code shorter so why not.