Sketch garbage collected transience ownership
The easiest way to create a global ID for owners is to allocate a
token in the heap.  For now, we do not do any reference counting.
This means that the transience ownership now only works with some form
of automatic tracing garbage collector, hence the name change.

It is tempting to just use the location of the owner as the token.
The problem with this approach is that a new owner may steal the
contents of something else.  For example:

```cpp
    auto v1 = immer::vector<int>{1, 2, 3}.transient();
    auto v2 = v1.persist();
    auto v3 = v1.transient();
    v3.set(1, 42);
    assert(v2[1] != 42)
```

The last assertion might fail, if `v3` happens to be constructed at
the same location where `v1` lived.  This would be vad bad :(  This is
why we are allocating the token, so the _id_ can not be reused as long
as there are references to it.

Note to self:

    An alternative approach to solve this problem would be to clean up
    the ownership tokens inside `id`.  This has a complexity
    proportional to the number of number of elements changed in the
    editing session times and the depth of the tree.