Add mechanism to test very small values for B
Interestingly most things even work for `B=1`, which forms a binary
tree.  `flex_vector` concatenation goes guacamole though, as expected,
because there is nothing to balance.  In fact, for `B=1`, optimal
concatenation can and should be done without relaxing the tree.  In
the other hand, `B=1` is a degenerate case that is only interesting as
a curiosity, so let's not really worry about it :D