Avoid push_tail step during concatenation
We change the concatenation operation so it knows how to deal with the
tail directly.  In this way, we save the `push_tail()` step before the
actual concatenation.

The solution is to pass the tail down the concatenation traversal.
When we are to concatenate the leaves (right-most leaf of the right
hand side, left-most leaf of the right hand side) by building a new
center node, we just put the tail in the middle of the new center
node.