Implement transient concatenation
Note that we do not do actual transient concatenation for arbitrary
ref-counted vectors.  It is very hard to do so with almost null
benefit, even more in an exception safe way.  We elaborate this
argument in the paper (hopefully to be published soon).

We still might add transient-move versions, because it might provide
some benefit when concatenating small vectors.