Add convinient constructors
It might make sense to optimize this a bit further.  In particular, we
can fill full buffers and unconditionally mutatingly push them.
Specially if we know the size of the input range, we can use some form
of "constructive recursion" to build the tree.

This is still faster than what most other implementations offer, so
whatever... :D