Increase the chances for short-circuiting equality check
Admittedly I am not completely happy with the solution.  The zipping
traversal is very heuristic and misses some potential short-circuits
while at the same time being kind of flaky.  I had a crash in Ewig
that I was not able to later reproduce that made me very suspicious,
but sadly I was not able to get any more crashes either using Ewig or
by writing convoluted tests.

The approach I would like most to for relaxed trees is similar to the
one based on iterators, but with a special iterator that keeps a stack
of the whole traversal.  When two leaf buffers are found to be
identical, we can traverse up the stack und jump the whole common
sub-tree.  This may also be implementable using
boost::coro2... performance doing per-item iteration was not so good,
but mauybe if we stop at an upper layer it is fine...