Change heap policy API to avoid space waste
We no longer create a free list unconditionally for the biggest type.
Instead, we now choose the target size (for example, the size of inner
nodes in a vector) and use the free list for the ones that are
actually smaller than that.  We can also just avoid a free list if the
type is of arbitrary size.

We might make further changes later, like, pick the inner node size of
vectors as universal optimized thingy and also optimize other
things (eg. boxed objects) that are withing a threshold to that.  But
I'd rather do these optimizations when we have more types that could
share free lists (maps!!!).