Make size_type std::size_t for vector types
We used std::uint32_t before because it achieves some marginal
performance gains for flex_vectors with lots of relaxed nodes.  This
also made the structure more comparable to that of librrb.

However, one very interesting thing that the fuzzer did show us it
that is easy to produce huge vectors by concatenating them to
themselves.  Because of structural sharing, this uses very little
memory indeed.  This might be an interesting way of building sparse
sequences.  Also, it makes sense to use > 4 million element vectors in
machines with lots of memory.  And of course, this makes things more
like the standard.  All in all, this seems like a good decission.

PS.  A follow up is needed to fix the descent positions accordingly.