Fix leaked relaxed node metadata when using transients
In this fuzzing session we also found ways to build too big vectors
that just produce shifts that overflow.  This could maybe be fixed by
checking for overflow in relaxed positions `index()` function, but
that is just a stupid pessimization for a degenerate case...  We added
assertions so a user that hits this could at least barely find what is
going on.