Fix copying transient should invalidate tokens
When we copy a transient we are just doing a shallow copy, thus
creating "aliases" of the nodes.  This means that the copied-from
object should not be able to mutate its recently created nodes
anymore, since they are now visible elsewhere.  Thus, we need to reset
its own token.

We also reset the token of the copied-into object.  This is because
otherwise we could also reach a similar situation by moving an object,
then copying into it.  This could be otherwise solved by "swapping"
the tokens during a move-assignment or move-construction, but in that
case we would need to make the move-constructor noexcept (to allocate
a new token for it).  With the current approach the moves are cheaper
and, more importantly, noexcept.

This commit fixes issue #16