Reduce the amount of overloading in rbts visitation to make MSVC happy
Ok.

So the problem is that we where really streasing the compiler by
having a huge overload set for the visit_node(visitor, ...) and
related set of functions.  Furthermore, these functions are variadic,
and some of the overloads also play SFINAE on their
arguments (arguments that make sense only for their own particular
kind of visitation: there are, for example, binary and unary visitors
under the same interface.)  MSVC did not interpret SFINAE propertly in
one instance where it was being passed an argument that did not make
sense and it was simply rising an error.

The newer approach, while less elegant for some narrow interpretations
of elegant (it requires a CRTP inheritance for the visitor, while the
previous approach might appeal better to some functional aesthetic
sensibilities) will probably make all compilers happier.  This means
lower compilation times, easier time optimizing, etc...