Do not set value of IMMER_{HAMTS,RBTS}_TAGGED_NODE if it is already set
Currently the data layout of the immer data types depends on the value
of the NDEBUG macro. This is rather unfortunate because if a library is
built with debugging disabled, and then linked into an application built
with debugging enabled, the code will crash with an assertion error.

Thus, I provide a way to explicitly set the layout intended for the
application so that it does not get overridden by the value of NDEBUG,
which allows the user to have a consistent data layout across multiple
code units with different values of debugging.