Limit the size of the fuzzer input
While we have inputs that are a bit larger like this in our bug
corpus, inputs larger than this have not really found interesting
bugs.

Given the structure of the fuzzers all interesting bugs must be found
with smaller inputs than this.  This should also improve performance
and allow searching this interesting space faster.