create a new memory policy: relocation_policy
This commit refactors the previous implementation of this PR so that
both the memory and compute overhead of tracking relocation information
in the iterator is only performed if a specific memory_policy is
enabled. We create a new relocation_policy policy and track two
different implementations: no_relocation_policy, which is the default in
all cases, and gc_relocation_policy, which should be enabled if the user
is using the library with a relocating garbage collector.

Because so much information about how to track relocation information is
type-specific, the policy acts merely as a marker, and each type that
needs to track relocation information is responsible for defining a
template that does nothing and has no members if relocation is disabled,
and otherwise stores the necessary information and maintains it if that
information is required. We then instrument the class that needs to
track data by inserting calls to methods on the relocation information
template, which will be responsible for updating the information as
necessary.