unit test for relocation info for champ_iterator
This test is slightly fragile in that it probably will not work if the
depth of the tree is uneven across different nodes. However, in order to
make it completely robust irrespective of the platform-specific details
of the hash function and the number of elements in the set being tested,
it would be necessary to re-implement a lot of the functionality for
walking the tree in the test, which seems counterproductive to me as we
would end up just testing the business logic against itself rather than
against a flat set of expectations.

So intead, I ended up just implementing a partial check for the
relocation information which tests the major behavior of the step_right
and step_down functions in the iterator, rather than testing the entire
set exhaustively. This allows the expected results of the test to be
specified declaratively, which makes the test more appropriate in terms
of what functionality it is actually testing.