Use sfinea to conditionally exclude iterator facade operators
This fixes an issue that showed up on recent compiler version
-- e.g. gcc 14.3 -- in which the std::sized_sentinel_for concept
would evaluate to true on forward iterator types in spite of the
fact that their operator- would fail to compile. This only occured
with C++20 or greater. This prevented doing things like

  auto m = immer::map<int,int>{{0,0}, {1,1}}
  auto v = std::vector<std::pair<int,int>>(m.begin(), m.end());

or

  auto n = std::ranges::distance(m.begin(), m.end());

This was due to the fact that static_asserts were used to exclude
certain operators at compile time, but C++ concepts basically ignore
those. This would then lead the standard library to chose
implementations relying on operators that couldn't compile.

Link:

https://en.cppreference.com/w/cpp/iterator/sized_sentinel_for.html