Fix free lists would misalign the data in a very stupid way
The problem was we were prepending the linked list data.  I don't know
why I was doing that to be honest, probably I just got a bit confused.
The thing is, this data is only needed when the object is deallocated,
so we can simply reuse the already allocated node instead of
prepending extra bits before.